+++ lib/MForm/Inputs/MFormInputsAbstract.php
@@ -42,9 +42,9 @@ abstract class MFormInputsAbstract
             $configForm->addMediaField($id.'bgImg', ['preview' => 1], null, ['label' => $this->config['bgImgLabel']]);
             $keys = array_diff($keys, ['bgImg']);
         }
-        if (in_array('bgColor', $keys) && isset($this->config['bgColor']) && is_array($this->config['bgColor']) && count($this->config['bgColor']) > 0) {
-            $configForm->addSelectField($id.'bgColor', $this->config['bgColor'], ['label' => $this->config['bgColorLabel']], 1, $this->config['bgColorDefaultValue']);
-            $keys = array_diff($keys, ['bgColor']);
+        if (in_array('bgClass', $keys) && isset($this->config['bgClass']) && is_array($this->config['bgClass']) && count($this->config['bgClass']) > 0) {
+            $configForm->addSelectField($id.'bgClass', $this->config['bgClass'], ['label' => $this->config['bgClassLabel']], 1, $this->config['bgClassDefaultValue']);
+            $keys = array_diff($keys, ['bgClass']);
         }
         foreach ($keys as $key) {
             if (!empty($this->config[$key]) && is_array($this->config[$key])) {