+++ lib/MForm/Repeater/MFormRepeaterHelper.php
@@ -136,6 +136,7 @@ class MFormRepeaterHelper
                 }
             } else if (($mformItem instanceof MForm && !isset($items[$key-1])) || ($mformItem instanceof MForm && (isset($items[$key-1]) && $items[$key-1] instanceof MFormItem && $items[$key-1]->getType() !== 'repeater'))) {
                 $obj = array_merge($obj, self::prepareChildItems($mformItem, $repeaterId, $group, $groups, $parentId));
+                self::addWidgetAttributesByMFormObj($mformItem, $repeaterId, $group, $groups, $parentId);
             } if ($mformItem instanceof MForm) {
                 foreach ($mformItem->getItems() as $item) {
                     if ($item instanceof MFormItem && $item->getType() === 'collapse') {
@@ -151,6 +152,15 @@ class MFormRepeaterHelper
     private static function getNameKey(MFormItem $mformItem): string
     {
         return ((is_array($mformItem->getVarId())) ? implode('.', $mformItem->getVarId()) : $mformItem->getVarId());
+    }
+
+    private static function addWidgetAttributesByMFormObj(MForm $mform, string $repeaterId, string $group, string $groups, string $parentId): void
+    {
+        foreach ($mform->getItems() as $item) {
+            if ($item instanceof MFormItem && $item->getType() === 'collapse') {
+                self::addWidgetAttributes($item, $repeaterId, $group, $groups, $parentId);
+            }
+        }
     }
 
     private static function addWidgetAttributes(MFormItem $mformItem, string $repeaterId, string $group, string $groups, string|null $parentId, string|null $nameKey = null): void
@@ -177,7 +187,7 @@ class MFormRepeaterHelper
         if (isset($mformItem->getAttributes()['data-toggle-item'])) {
             $mformItem->addAttribute(':data-toggle-item', "'" . $mformItem->getAttributes()['data-toggle-item'] . "-'+" . $repeaterId . "Index" . ((!empty($parentId)) ? "+'-'+" . $parentId . 'Index' : ''));
         }
-        if ($mformItem->getType() === 'collapse' && isset($mformItem->getAttributes()['data-group-collapse-id'])) {
+        if (isset($mformItem->getAttributes()['data-group-collapse-id'])) {
             $mformItem->addAttribute(':data-group-collapse-id', "'" . $mformItem->getAttributes()['data-group-collapse-id'] . "-'+" . $repeaterId . "Index" . ((!empty($parentId)) ? "+'-'+" . $parentId . 'Index' : ''));
         }
     }